Moved traits related to send packets in their own mod
