+++ crates/relayer-components/src/chain/traits/queries/mod.rs
@@ -3,6 +3,7 @@ pub mod connection;
 pub mod consensus_state;
 pub mod packet_commitments;
 pub mod received_packet;
+pub mod send_packet;
 pub mod status;
 pub mod unreceived_packets;
 pub mod write_ack;
