+++ crates/relayer-components/src/chain/traits/queries/unreceived_packets.rs
@@ -21,47 +21,3 @@ where
         sequences: &[Counterparty::Sequence],
     ) -> Result<Vec<Counterparty::Sequence>, Self::Error>;
 }
-
-#[async_trait]
-pub trait CanQuerySendPacketFromSequence<Counterparty>:
-    HasIbcChainTypes<Counterparty> + HasIbcPacketTypes<Counterparty> + HasErrorType
-where
-    Counterparty: HasIbcChainTypes<Self>,
-{
-    /// Given a single sequence, a channel and port will query the outgoing
-    /// packets if it hasn't been relayed.
-    async fn query_send_packet_from_sequence(
-        &self,
-        channel_id: &Self::ChannelId,
-        port_id: &Self::PortId,
-        counterparty_channel_id: &Counterparty::ChannelId,
-        counterparty_port_id: &Counterparty::PortId,
-        sequence: &Self::Sequence,
-        // The height is given to query the packets from a specific height.
-        // This height should be the same as the query height from the
-        // `CanQueryPacketCommitments` made on the same chain.
-        height: &Self::Height,
-    ) -> Result<Self::OutgoingPacket, Self::Error>;
-}
-
-#[async_trait]
-pub trait CanQuerySendPacketsFromSequences<Counterparty>:
-    HasIbcChainTypes<Counterparty> + HasIbcPacketTypes<Counterparty> + HasErrorType
-where
-    Counterparty: HasIbcChainTypes<Self>,
-{
-    /// Given a list of sequences, a channel and port will query a list of outgoing
-    /// packets which have not been relayed.
-    async fn query_send_packets_from_sequences(
-        &self,
-        channel_id: &Self::ChannelId,
-        port_id: &Self::PortId,
-        counterparty_channel_id: &Counterparty::ChannelId,
-        counterparty_port_id: &Counterparty::PortId,
-        sequences: &[Self::Sequence],
-        // The height is given to query the packets from a specific height.
-        // This height should be the same as the query height from the
-        // `CanQueryPacketCommitments` made on the same chain.
-        height: &Self::Height,
-    ) -> Result<Vec<Self::OutgoingPacket>, Self::Error>;
-}
